Fixed format of spark parameters

+++ atpipeline/at_spark.py
@@ -17,10 +17,10 @@ class Spark:
         self.default_parallelism = 2 * hostCores
 
         #The driver memory can be set to half of the system memory
-        self.driver_memory       = hostMemory / 2.0
+        self.driver_memory       = int(hostMemory / 2.0)
 
         #The driver memory can be set to half of the system memory, and increased if the job fails
-        self.executor_memory     = hostMemory / 2.0
+        self.executor_memory     = int(hostMemory / 2.0)
 
         #According to Gayathri and FC, executor_cores is some default constant, k. k == 4 according to FC
         self.executor_cores      = 4
